# Improve recursion performance v1