--- a/Jint.Tests/Runtime/EngineLimitTests.cs
+++ b/Jint.Tests/Runtime/EngineLimitTests.cs
@@ -7,20 +7,24 @@ public class EngineLimitTests
     [Fact]
     public void ShouldAllowReasonableCallStackDepth()
     {
-        // 249
-        var functionNestingCount = 300;
+#if RELEASE
+        const int FunctionNestingCount = 350;
+#else
+        const int FunctionNestingCount = 170;
+#endif
+
         // generate call tree
         var sb = new StringBuilder();
         sb.AppendLine("var x = 10;");
         sb.AppendLine();
-        for (var i = 1; i <= functionNestingCount; ++i)
+        for (var i = 1; i <= FunctionNestingCount; ++i)
         {
             sb.Append("function func").Append(i).Append("(func").Append(i).AppendLine("Param) {");
             sb.Append("    ");
-            if (i != functionNestingCount)
+            if (i != FunctionNestingCount)
             {
                 // just to create a bit more nesting add some constructs
-                sb.Append("try { return x++ > 1 ? func").Append(i + 1).Append("(func").Append(i).AppendLine("Param): undefined; } finally { /* nothing*/ }");
+                sb.Append("return x++ > 1 ? func").Append(i + 1).Append("(func").Append(i).AppendLine("Param): undefined;");
             }
             else
             {
@@ -36,7 +40,4 @@ public class EngineLimitTests
         engine.Execute(sb.ToString());
         Assert.Equal(123, engine.Evaluate("func1(123);").AsNumber());
     }
-
-    public static IEnumerable<object[]> Data =>
-        Enumerable.Range(380, 10).Select(x => new object[] { x });
 }