--- a/Jint/Runtime/Interpreter/Expressions/JintCallExpression.cs
+++ b/Jint/Runtime/Interpreter/Expressions/JintCallExpression.cs
@@ -1,3 +1,5 @@
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.CompilerServices;
 using Esprima.Ast;
 using Jint.Native;
 using Jint.Native.Function;
@@ -69,10 +71,135 @@ namespace Jint.Runtime.Interpreter.Expressions
 
         protected override ExpressionResult EvaluateInternal(EvaluationContext context)
         {
-            return NormalCompletion(_calleeExpression is JintSuperExpression
-                ? SuperCall(context)
-                : Call(context)
-            );
+            if (_calleeExpression._expression.Type == Nodes.Super)
+            {
+                return NormalCompletion(SuperCall(context));
+            }
+
+            // https://tc39.es/ecma262/#sec-function-calls
+
+            var reference = _calleeExpression.Evaluate(context).Value;
+
+            if (ReferenceEquals(reference, Undefined.Instance))
+            {
+                return NormalCompletion(Undefined.Instance);
+            }
+
+            var engine = context.Engine;
+            var func = engine.GetValue(reference, false);
+
+            if (func.IsNullOrUndefined() && _expression.IsOptional())
+            {
+                return NormalCompletion(Undefined.Instance);
+            }
+
+            if (reference is Reference referenceRecord
+                && !referenceRecord.IsPropertyReference()
+                && referenceRecord.GetReferencedName() == CommonProperties.Eval
+                && ReferenceEquals(func, engine.Realm.Intrinsics.Eval))
+            {
+                return HandleEval(context, func, engine, referenceRecord);
+            }
+
+            var thisCall = (CallExpression) _expression;
+            var tailCall = IsInTailPosition(thisCall);
+
+            // https://tc39.es/ecma262/#sec-evaluatecall
+
+            JsValue thisValue;
+            var referenceRecord1 = reference as Reference;
+            if (referenceRecord1 is not null)
+            {
+                if (referenceRecord1.IsPropertyReference())
+                {
+                    thisValue = referenceRecord1.GetThisValue();
+                }
+                else
+                {
+                    var baseValue = referenceRecord1.GetBase();
+
+                    // deviation from the spec to support null-propagation helper
+                    if (baseValue.IsNullOrUndefined()
+                        && engine._referenceResolver.TryUnresolvableReference(engine, referenceRecord1, out var value))
+                    {
+                        thisValue = value;
+                    }
+                    else
+                    {
+                        var refEnv = (EnvironmentRecord) baseValue;
+                        thisValue = refEnv.WithBaseObject();
+                    }
+                }
+            }
+            else
+            {
+                thisValue = Undefined.Instance;
+            }
+
+            var argList = ArgumentListEvaluation(context);
+
+            if (!func.IsObject() && !engine._referenceResolver.TryGetCallable(engine, reference, out func))
+            {
+                ThrowMemberisNotFunction(referenceRecord1, reference, engine);
+            }
+
+            var callable = func as ICallable;
+            if (callable is null)
+            {
+                ThrowReferenceNotFunction(referenceRecord1, reference, engine);
+            }
+
+            if (tailCall)
+            {
+                // TODO tail call
+                // PrepareForTailCall();
+            }
+
+            var result = engine.Call(callable, thisValue, argList, _calleeExpression);
+
+            if (!_cached && argList.Length > 0)
+            {
+                engine._jsValueArrayPool.ReturnArray(argList);
+            }
+
+            engine._referencePool.Return(referenceRecord1);
+            return NormalCompletion(result);
+        }
+
+        [DoesNotReturn]
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private static void ThrowReferenceNotFunction(Reference? referenceRecord1, object reference, Engine engine)
+        {
+            var message = $"{referenceRecord1?.GetReferencedName() ?? reference} is not a function";
+            ExceptionHelper.ThrowTypeError(engine.Realm, message);
+        }
+
+        [DoesNotReturn]
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private static void ThrowMemberisNotFunction(Reference? referenceRecord1, object reference, Engine engine)
+        {
+            var message = referenceRecord1 == null
+                ? reference + " is not a function"
+                : $"Property '{referenceRecord1.GetReferencedName()}' of object is not a function";
+            ExceptionHelper.ThrowTypeError(engine.Realm, message);
+        }
+
+        private ExpressionResult HandleEval(EvaluationContext context, JsValue func, Engine engine, Reference referenceRecord)
+        {
+            var argList = ArgumentListEvaluation(context);
+            if (argList.Length == 0)
+            {
+                return NormalCompletion(Undefined.Instance);
+            }
+
+            var evalFunctionInstance = (EvalFunctionInstance) func;
+            var evalArg = argList[0];
+            var strictCaller = StrictModeScope.IsStrictModeCode;
+            var evalRealm = evalFunctionInstance._realm;
+            var direct = !_expression.IsOptional();
+            var value = evalFunctionInstance.PerformEval(evalArg, evalRealm, strictCaller, direct);
+            engine._referencePool.Return(referenceRecord);
+            return NormalCompletion(value);
         }
 
         private JsValue SuperCall(EvaluationContext context)
@@ -101,122 +228,6 @@ namespace Jint.Runtime.Interpreter.Expressions
             var activeFunction = envRec._functionObject;
             var superConstructor = activeFunction.GetPrototypeOf();
             return superConstructor;
-        }
-
-        /// <summary>
-        /// https://tc39.es/ecma262/#sec-function-calls
-        /// </summary>
-        private JsValue Call(EvaluationContext context)
-        {
-            var reference = _calleeExpression.Evaluate(context).Value;
-
-            if (ReferenceEquals(reference, Undefined.Instance))
-            {
-                return Undefined.Instance;
-            }
-
-            var engine = context.Engine;
-            var func = engine.GetValue(reference, false);
-
-            if (func.IsNullOrUndefined() && _expression.IsOptional())
-            {
-                return Undefined.Instance;
-            }
-
-            if (reference is Reference referenceRecord
-                && !referenceRecord.IsPropertyReference()
-                && referenceRecord.GetReferencedName() == CommonProperties.Eval
-                && ReferenceEquals(func, engine.Realm.Intrinsics.Eval))
-            {
-                var argList = ArgumentListEvaluation(context);
-                if (argList.Length == 0)
-                {
-                    return Undefined.Instance;
-                }
-
-                var evalFunctionInstance = (EvalFunctionInstance) func;
-                var evalArg = argList[0];
-                var strictCaller = StrictModeScope.IsStrictModeCode;
-                var evalRealm = evalFunctionInstance._realm;
-                var direct = !_expression.IsOptional();
-                var value = evalFunctionInstance.PerformEval(evalArg, evalRealm, strictCaller, direct);
-                engine._referencePool.Return(referenceRecord);
-                return value;
-            }
-
-            var thisCall = (CallExpression) _expression;
-            var tailCall = IsInTailPosition(thisCall);
-            return EvaluateCall(context, func, reference, thisCall.Arguments, tailCall);
-        }
-
-        /// <summary>
-        /// https://tc39.es/ecma262/#sec-evaluatecall
-        /// </summary>
-        private JsValue EvaluateCall(EvaluationContext context, JsValue func, object reference, in NodeList<Expression> arguments, bool tailPosition)
-        {
-            JsValue thisValue;
-            var referenceRecord = reference as Reference;
-            var engine = context.Engine;
-            if (referenceRecord is not null)
-            {
-                if (referenceRecord.IsPropertyReference())
-                {
-                    thisValue = referenceRecord.GetThisValue();
-                }
-                else
-                {
-                    var baseValue = referenceRecord.GetBase();
-
-                    // deviation from the spec to support null-propagation helper
-                    if (baseValue.IsNullOrUndefined()
-                        && engine._referenceResolver.TryUnresolvableReference(engine, referenceRecord, out var value))
-                    {
-                        thisValue = value;
-                    }
-                    else
-                    {
-                        var refEnv = (EnvironmentRecord) baseValue;
-                        thisValue = refEnv.WithBaseObject();
-                    }
-                }
-            }
-            else
-            {
-                thisValue = Undefined.Instance;
-            }
-
-            var argList = ArgumentListEvaluation(context);
-
-            if (!func.IsObject() && !engine._referenceResolver.TryGetCallable(engine, reference, out func))
-            {
-                var message = referenceRecord == null
-                    ? reference + " is not a function"
-                    : $"Property '{referenceRecord.GetReferencedName()}' of object is not a function";
-                ExceptionHelper.ThrowTypeError(engine.Realm, message);
-            }
-
-            var callable = func as ICallable;
-            if (callable is null)
-            {
-                var message = $"{referenceRecord?.GetReferencedName() ?? reference} is not a function";
-                ExceptionHelper.ThrowTypeError(engine.Realm, message);
-            }
-
-            if (tailPosition)
-            {
-                // TODO tail call
-                // PrepareForTailCall();
-            }
-
-            var result = engine.Call(callable, thisValue, argList, _calleeExpression);
-
-            if (!_cached && argList.Length > 0)
-            {
-                engine._jsValueArrayPool.ReturnArray(argList);
-            }
-
-            engine._referencePool.Return(referenceRecord);
-            return result;
         }
 
         /// <summary>